+++ app/Http/Controllers/Backsite/SpecialistController.php
@@ -41,7 +41,7 @@ class SpecialistController extends Controller
      */
     public function index()
     {
-        // abort_if(Gate::denies('specialist_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('specialist_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $specialist = Specialist::orderBy('created_at', 'desc')->get();
 
@@ -88,7 +88,7 @@ class SpecialistController extends Controller
      */
     public function show(Specialist $specialist)
     {
-        // abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('pages.backsite.master-data.specialist.show', compact('specialist'));
     }
@@ -101,7 +101,7 @@ class SpecialistController extends Controller
      */
     public function edit(Specialist $specialist)
     {
-        // abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return view('pages.backsite.master-data.specialist.edit', compact('specialist'));
     }
@@ -137,7 +137,7 @@ class SpecialistController extends Controller
      */
     public function destroy(Specialist $specialist)
     {
-        // abort_if(Gate::denies('specialist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('specialist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $specialist->forceDelete();
